grooming & ketchup; rlvex
//3.959 Medea/Lucian: g_iAuth was not being set in menu path to adding
user, resulting in no auth value for return menu. Added  g_iAuth=auth;
to fetchAvi command.

//3.936 Bugfixes:  1. Reported issue, message about UUID not found when
trying to add other. I haven't found this to be repeatable, but I have
managed to get it to do it once. Increased llRequestAgentInfo timer to 4
seconds from 0.5, because 0.5 may simply be too short to reliably get a
response from the dataserver, particularly in a laggy sim. Hopefully
this will remove the issue, but let's monitor this.  2. Reported issue
in secowner exceptions not being set. Typo in settings was looking for
auth_secowner rather than auth_secowners in LM_SETTINGS_RESPONSE, so
secowners list was not getting filled.

//3.934 notes. I've overhauled the menu appearance as it wasn't really
that easy to understand, following the RLV terminology which is a bit of
a headache for scripters, let alone users. Descriptions are clearer and
now represent the current state, and we use unicode ☐ and ☒ to indicate
selected/deselected state on the buttons, common with unicode
enhancements to other menus. ALL buttons indicators swapped to make
clearer the effect of pressing them, and the person to whom the
exceptions will be added/removed is now named at the top of the menu.
Finally, I've cleared up remenuing a little. When someone who is not an
owner/secowner is set to default and removed from the list, we now
return the personlist, and when a new person is added, we remenu. -
Medea

//3.934 notes. Removed SetOwnerExs as it's never called. Added a check
to setAllExs to  avoid setting exceptions for wearer (for example if the
wearer is set as an owner) as they're pointless. - Medea