add ice cream
This adds a global string for ice cream. That's right! We can now rename
our "vanilla" button to any flavour of ice cream the user likes best.
Just in case, the button stays at the same spot regardless, even with a
different name and we also retain the old "vanilla on/off" command as a
fall-back if somebody forget their favourite flavour! :speak_no_evil: