remove emergency news logic
Settings was a good place for this feature but there is no reason to
keep it there if we summarize this type of logic (hopefully) in a single
web connect script.

We never needed to use the emergency news in 6.x, and if I recall
correctly then the last time we used it was sometime around 3.844 when
the name2key queries were failing in early 2013.

#1032