prepare for oc_root (installer)
We have to make sure to not end up with duplicated functionality. This
will be ultimately the installer's job but it would be very frustrating
for users who prefer manual refreshes if they first need to figure out
why everything happens twice and in a different way. :speak_no_evil: