handle semi-dynamic menus and rebooting
We register the Main, Apps and Settings menus and query for the
Animations, RLV and Kidnap menus. To help the user, we take care of this
in a semi-dynamic style for easier recognition. The Settings menu needs
to be able to handle adding and removing trigger buttons/menus for the
oc_resizer plugin as some users will want to remove this plugin after
they are done adjusting their device.

When the script has a new owner (wearer) we have to reset, so handling
the above has to happen then and also when we start a new session.
(Maybe we should have a function that takes care of the timer event.)

For ease of use and compatibility we have a command "reboot" to reset
the script.