add several functions and make use of oc_dialog
To make use of oc_dialog we need a function and some logic for dialog
responses in our link message event. Then we add functions for our
menus, root (main), apps, settings and about and related response
handling.

The main menu is "semi-dynamic", meaning that buttons won't jumble
around alphabetically but if a menu/feature isn't there, we have an
empty socket [  -  ] instead. This is done for ease of recognizability.
(We will code the other menu functions later.)

We also want a (chat) commands function and need a call for that in the
link message event, and create another function init() for our timer
event (this function should come in handy for some more stuff yet).

The reboot works via chat command now. Let's also do a reboot when we
change owner (wearer).

PS: There is also contemplation about whether to add logic for the lock
here or to use a dedicated plugin, same for version control and
updating. I'm tending towards the latter mostly because if you make your
own installer, it is easy to swap a couple plugins with yours, but it is
cumbersome to diff and manually merge changes each time the upstream
releases new code. Let's see...