code other menus, add settings response and token
We have to make sure that existing behavior doesn't change too much (at
least in our version of the new source). There is the "hide all" feature
in the settings menu aka the "stealth toggle". We do it right from the
beginning this time and reflect the state of stealth correctly, add a
variable for that, check when relevant.

We also want to use a dedicated lock plugin (not yet coded, just use a
dummy response if you follow my progress live) and need to learn about
the state of the lock.

Then we also need a simple function for nameURI for the 4.0/Six hotlink
style in menus and prints. That comes in play for a variable (dist) that
we receive in the settings response event, which we need for a properly
formatted name of origin.

The about menu shows version (natural) number and even the origin. We
have to ensure that receiving users are informed that our project can't
take the burden of support in case the source gets distributed by
others. :speak_no_evil:

The about menu is just a plain info "pop-up" where we don't need to make
use of the dialog function but the buttons for app and settings menu
functions need to be handled in the dialog response, so we can call up
the menus and features (list of apps, print/load/fix, stealth toggle,
adjuster menus etc..)

PS: Also added urls to our manual pages in the headlines, format:
[url<space>name]