add variable for safeword and familiar commands
We want chat commands for passing users (and their authorized partners)
a help or a license file. We also need to handle our "fix menus" command
and one for printing general info (about) the device. Some users will be
familiar with either "version" or "info" so let's let both do the same
thing in this new source. That feature is also often used to look up the
safeword, so we need a variable (default is RED) for that and get it in
the settings response event. We also should make use of a failsafe to
avoid accidental script duplicates and to play nice with updaters.