add update function and confirmation menu
We generate a random remote access pin and announce to installers that
we are ready when called. If we are running packages from the upstream
we use that channel, else we use a different one. We let the user
confirm or cancel before an installation starts.

PS: This is a very simple approach to updating and I believe that it is
more useful to everyone when easier to understand and modify. There are
no chat commands and no tests if patches are deprecated. Users can
choose whether to run patches from the upstream or not. Installations
are triggered by touching the installer.