merge peanut branch
This is the full merge of the version "Peanut" branch. When using a diff
tool it should be pretty clear what happens in detail in each script,
but I will try to summarize this large change-set real quick for your
convenience:

Copyright notices are conclusive with date ranges, authors and source
links, even if you have to scroll a bit now. These are how they have to
be when taking into account that users are passing LSL scripts around as
single files, and usually without any license included.

I use build numbers as individual versions for each script, and have
been doing so since a long time on my personal branches. This allows us
to distinguish which is which once the scripts are out in the wild, and
now also without having to open the collar and looking. A command
"<prefix> list builds" can be used to print all build numbers to the
chat history.

The so-called "schmoDialog" has always been an issue, I've tried with
the best of my ability to optimize this program, and believe that I
finally could with the help of @GarvinTwine. A lot of its annoying
antics should be gone now.

Optimization happened all over and the scripts are less encumbered by
ancient commentary that was misleading and no longer current. The
scripts are not the place to document the code, for that purpose we use
git version control since 2011. There is much improvement to overall
readability now.

There are also two new scripts, one is a variation of themes that puts
looks first. This is for users of legacy artwork because it resembles
the old "Appearance" bundles more closely right out of the box. The
other new script is a reduction of the couples program to only the
non-looped "pet" sequence.

The root script does no longer have update logic in it, but instead we
have the optional update script that I wrote last year. For most
builders this should make it a lot more simple to decided whether they
want to allow future updates for their collars or not. I still think it
would be clearer if the root script was separate from the stealth and
lock logic, but user feedback wanted it to be all combined.

Finally, I have replaced the AntiSlide logic with a walk overrider that
plays a puppy crawl when the collar wearer moves and is also in a collar
pose at the same time.

NOTE: The products that are distributed by virtualdisgrace.com on Second
Life® have refreshed pose sets by Ash Ellsberg, as well as a new
particle chain that I based on original artwork by @GreyMars.