rebrand :v:
This commit rebrands the entirety of Wendy's OpenCollar Distribution to
become the Peanut Collar Distribution. While my authorship will remain
forever in the various blends of OpenCollar, I've decided that I want to
sever any remaining ties to what, at the time of this writing, is being
disputed to become the "official" OpenCollar project. I want to distance
myself and virtualdisgrace.com from the OpenCollar name and what it
currently represents as far as possible. What my friends at
virtualdisgrace.com and I contributed to was intended to be a fun toy
for everyone, regardless of their heritage, religion, their chosen
gender, and sexual orientation. I want that Peanut can continue to be
the weirdo toy for our quirky-cute community. It was that diversity and
inclusiveness that motivated us so much in the years 2011 to 2018, when
we all suddenly decided that it would be a great idea to nerdgasm over
wearing dog collar gadgets. We had so much fun then, and nothing is
going to keep us from having it also today! This is what OUR name stands
for! :heart: :peace_symbol: :rainbow_flag: